Refactor to merge projects
