+++ tests/pages/mobile/auth_page.py
