+++ tests/pages/mobile/photo_page.py
