+++ tests/pages/mobile/user_add_album_photo_page.py
