+++ tests/pages/mobile/user_album_edit_page.py
