+++ tests/pages/mobile/user_album_page.py
