+++ tests/pages/mobile/user_albums_page.py
