+++ tests/pages/mobile/user_edit_album_photo_page.py
